Moved OnDeployScriptBase out of impl so that it can be extended by other OSGi bundles
