+++ bundle/src/test/java/com/adobe/acs/commons/ondeploy/scripts/OnDeployScriptTestExampleFailCtor.java
@@ -1,4 +1,6 @@
-package com.adobe.acs.commons.ondeploy.impl;
+package com.adobe.acs.commons.ondeploy.scripts;
+
+import com.adobe.acs.commons.ondeploy.scripts.OnDeployScriptBase;
 
 public class OnDeployScriptTestExampleFailCtor extends OnDeployScriptBase {
     private OnDeployScriptTestExampleFailCtor() {
