+++ bundle/src/test/java/com/adobe/acs/commons/ondeploy/scripts/OnDeployScriptTestExampleSuccess1.java
@@ -1,4 +1,6 @@
-package com.adobe.acs.commons.ondeploy.impl;
+package com.adobe.acs.commons.ondeploy.scripts;
+
+import com.adobe.acs.commons.ondeploy.scripts.OnDeployScriptBase;
 
 public class OnDeployScriptTestExampleSuccess1 extends OnDeployScriptBase {
     @Override
